+++ app/services/update_user.rb
@@ -1,10 +1,11 @@
 class UpdateUser
   attr_accessor :user, :organisation, :reset_mfa
 
-  def initialize(user:, organisation:, reset_mfa: false)
+  def initialize(user:, organisation:, active: true, reset_mfa: false)
     self.user = user
     self.organisation = organisation
     self.reset_mfa = reset_mfa
+    @active = active
   end
 
   def call
@@ -17,6 +18,8 @@ class UpdateUser
         user.mobile_number = nil
         user.mobile_number_confirmed_at = nil
       end
+
+      user.deactivated_at = @active ? nil : DateTime.now
 
       user.save
     end